LayoutTests/imported/w3c/IndexedDB-private-browsing/idbdatabase_transaction5-expected.txt


PASS IDBDatabase.transaction() - If storeNames is an empty list, the implementation must throw a DOMException of type InvalidAccessError
